backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
python-dotenv
jinja2
passlib[bcrypt]
python-jose
openai
faiss-cpu
numpy
